requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
email-validator
python-multipart